An option on an index, which settles in cash because an index cannot be delivered

It adds a guarantee to Option rather than members: the option is in an index options segment, so its default underlying is the index, found by symbol as a NonTradeableInstrument unless it is given or UBI links it. The family classes such as EquityIndexOption inherit it.

Super classes

PriceAnalysis -> PriceStatistics -> OverlapStudies -> MomentumIndicators -> VolumeIndicators -> CycleIndicators -> PriceTransforms -> VolatilityIndicators -> StatisticFunctions -> MathTransforms -> MathOperators -> CandlestickPatterns -> Signals -> StrategyBacktests -> PerformanceMeasures -> Instrument -> TradeableInstrument -> Derivative -> Option -> IndexOption

Methods

Public methods

IndexOption$new()

IndexOption$clone()

Inherited methods

PriceStatistics$price_high()

PriceStatistics$price_histogram()

PriceStatistics$price_kurtosis()

PriceStatistics$price_low()

PriceStatistics$price_mean()

PriceStatistics$price_mean_absolute_deviation()

PriceStatistics$price_median()

PriceStatistics$price_quantile()

PriceStatistics$price_skewness()

PriceStatistics$price_standard_deviation()

PriceStatistics$price_summary()

PriceStatistics$price_variance()

PriceStatistics$returns()

PriceStatistics$returns_high()

PriceStatistics$returns_histogram()

PriceStatistics$returns_kurtosis()

PriceStatistics$returns_low()

PriceStatistics$returns_mean()

PriceStatistics$returns_mean_absolute_deviation()

PriceStatistics$returns_median()

PriceStatistics$returns_quantile()

PriceStatistics$returns_skewness()

PriceStatistics$returns_standard_deviation()

PriceStatistics$returns_summary()

PriceStatistics$returns_variance()

PriceStatistics$volume_high()

PriceStatistics$volume_histogram()

PriceStatistics$volume_kurtosis()

PriceStatistics$volume_low()

PriceStatistics$volume_mean()

PriceStatistics$volume_mean_absolute_deviation()

PriceStatistics$volume_median()

PriceStatistics$volume_quantile()

PriceStatistics$volume_skewness()

PriceStatistics$volume_standard_deviation()

PriceStatistics$volume_summary()

PriceStatistics$volume_total()

PriceStatistics$volume_variance()

PriceStatistics$volumes()

OverlapStudies$bollinger_bands()

OverlapStudies$double_exponential_moving_average()

OverlapStudies$exponential_moving_average()

OverlapStudies$kaufman_adaptive_moving_average()

OverlapStudies$mesa_adaptive_moving_average()

OverlapStudies$mid_point()

OverlapStudies$middle_price()

OverlapStudies$parabolic_sar()

OverlapStudies$simple_moving_average()

OverlapStudies$triangular_moving_average()

OverlapStudies$triple_exponential_moving_average()

OverlapStudies$weighted_moving_average()

MomentumIndicators$absolute_price_oscillator()

MomentumIndicators$aroon()

MomentumIndicators$aroon_oscillator()

MomentumIndicators$average_directional_movement_index()

MomentumIndicators$average_directional_movement_index_rating()

MomentumIndicators$balance_of_power()

MomentumIndicators$chande_momentum_oscillator()

MomentumIndicators$commodity_channel_index()

MomentumIndicators$directional_movement_index()

MomentumIndicators$minus_directional_indicator()

MomentumIndicators$minus_directional_movement()

MomentumIndicators$momentum()

MomentumIndicators$money_flow_index()

MomentumIndicators$moving_average_convergence_divergence()

MomentumIndicators$moving_average_convergence_divergence_extended()

MomentumIndicators$percentage_price_oscillator()

MomentumIndicators$plus_directional_indicator()

MomentumIndicators$plus_directional_movement()

MomentumIndicators$rate_of_change()

MomentumIndicators$rate_of_change_percent()

MomentumIndicators$rate_of_change_ratio()

MomentumIndicators$relative_strength_index()

MomentumIndicators$stochastic_fast_oscillator()

MomentumIndicators$stochastic_oscillator()

MomentumIndicators$stochastic_relative_strength_index()

MomentumIndicators$trix()

MomentumIndicators$ultimate_oscillator()

MomentumIndicators$williams_percent_r()

VolumeIndicators$chaikin_accumulation_distribution_line()

VolumeIndicators$chaikin_accumulation_distribution_oscillator()

VolumeIndicators$on_balance_volume()

CycleIndicators$hilbert_transform_dominant_cycle_period()

CycleIndicators$hilbert_transform_dominant_cycle_phase()

CycleIndicators$hilbert_transform_phasor_components()

CycleIndicators$hilbert_transform_sine_wave()

CycleIndicators$hilbert_transform_trend_line()

CycleIndicators$hilbert_transform_trend_mode()

PriceTransforms$average_price()

PriceTransforms$median_price()

PriceTransforms$typical_price()

PriceTransforms$weighted_close()

VolatilityIndicators$average_true_range()

VolatilityIndicators$normalized_average_true_range()

VolatilityIndicators$true_range()

StatisticFunctions$beta()

StatisticFunctions$correlation_coefficient()

StatisticFunctions$linear_regression()

StatisticFunctions$linear_regression_angle()

StatisticFunctions$linear_regression_intercept()

StatisticFunctions$linear_regression_slope()

StatisticFunctions$standard_deviation()

StatisticFunctions$variance()

MathTransforms$arc_cosine()

MathTransforms$arc_sine()

MathTransforms$arc_tangent()

MathTransforms$ceiling()

MathTransforms$cosine()

MathTransforms$exponential()

MathTransforms$floor()

MathTransforms$hyperbolic_cosine()

MathTransforms$hyperbolic_sine()

MathTransforms$hyperbolic_tangent()

MathTransforms$logarithm_base_10()

MathTransforms$natural_logarithm()

MathTransforms$sine()

MathTransforms$square_root()

MathTransforms$tangent()

MathOperators$add()

MathOperators$divide()

MathOperators$maximum()

MathOperators$maximum_index()

MathOperators$minimum()

MathOperators$minimum_index()

MathOperators$minimum_maximum()

MathOperators$minimum_maximum_index()

MathOperators$multiply()

MathOperators$subtract()

CandlestickPatterns$candle_abandoned_baby()

CandlestickPatterns$candle_advance_block()

CandlestickPatterns$candle_belt_hold()

CandlestickPatterns$candle_breakaway()

CandlestickPatterns$candle_closing_marubozu()

CandlestickPatterns$candle_concealing_baby_swallow()

CandlestickPatterns$candle_counter_attack()

CandlestickPatterns$candle_dark_cloud_cover()

CandlestickPatterns$candle_doji()

CandlestickPatterns$candle_doji_star()

CandlestickPatterns$candle_dragonfly_doji()

CandlestickPatterns$candle_engulfing()

CandlestickPatterns$candle_evening_doji_star()

CandlestickPatterns$candle_evening_star()

CandlestickPatterns$candle_gravestone_doji()

CandlestickPatterns$candle_hammer()

CandlestickPatterns$candle_hanging_man()

CandlestickPatterns$candle_harami()

CandlestickPatterns$candle_harami_cross()

CandlestickPatterns$candle_high_wave()

CandlestickPatterns$candle_hikkake()

CandlestickPatterns$candle_homing_pigeon()

CandlestickPatterns$candle_identical_three_crows()

CandlestickPatterns$candle_in_neck()

CandlestickPatterns$candle_inverted_hammer()

CandlestickPatterns$candle_kicking()

CandlestickPatterns$candle_kicking_by_length()

CandlestickPatterns$candle_ladder_bottom()

CandlestickPatterns$candle_long_legged_doji()

CandlestickPatterns$candle_long_line()

CandlestickPatterns$candle_marubozu()

CandlestickPatterns$candle_mat_hold()

CandlestickPatterns$candle_matching_low()

CandlestickPatterns$candle_modified_hikkake()

CandlestickPatterns$candle_morning_star()

CandlestickPatterns$candle_morning_star_doji()

CandlestickPatterns$candle_on_neck()

CandlestickPatterns$candle_piercing()

CandlestickPatterns$candle_rickshaw_man()

CandlestickPatterns$candle_rise_fall_three_methods()

CandlestickPatterns$candle_separating_lines()

CandlestickPatterns$candle_shooting_star()

CandlestickPatterns$candle_short_line()

CandlestickPatterns$candle_side_by_side_white_lines()

CandlestickPatterns$candle_spinning_top()

CandlestickPatterns$candle_stalled_pattern()

CandlestickPatterns$candle_stick_sandwich()

CandlestickPatterns$candle_takuri()

CandlestickPatterns$candle_tasuki_gap()

CandlestickPatterns$candle_three_black_crows()

CandlestickPatterns$candle_three_inside_up_down()

CandlestickPatterns$candle_three_line_strike()

CandlestickPatterns$candle_three_outside_up_down()

CandlestickPatterns$candle_three_stars_in_the_south()

CandlestickPatterns$candle_three_white_soldiers()

CandlestickPatterns$candle_thrusting_pattern()

CandlestickPatterns$candle_tristar()

CandlestickPatterns$candle_two_crows()

CandlestickPatterns$candle_unique_three_river()

CandlestickPatterns$candle_up_side_down_side_gap_three_methods()

CandlestickPatterns$candle_up_side_gap_two_crows()

Signals$is_cross_over()

Signals$is_cross_under()

StrategyBacktests$run_backtest()

PerformanceMeasures$alpha()

PerformanceMeasures$annualised_return()

PerformanceMeasures$annualised_volatility()

PerformanceMeasures$benchmark_beta()

PerformanceMeasures$calmar_ratio()

PerformanceMeasures$cumulative_return()

PerformanceMeasures$down_capture_ratio()

PerformanceMeasures$drawdowns()

PerformanceMeasures$expected_shortfall()

PerformanceMeasures$information_ratio()

PerformanceMeasures$maximum_drawdown()

PerformanceMeasures$performance_summary()

PerformanceMeasures$sharpe_ratio()

PerformanceMeasures$sortino_ratio()

PerformanceMeasures$tracking_error()

PerformanceMeasures$up_capture_ratio()

PerformanceMeasures$value_at_risk()

Instrument$equals()

Instrument$format()

Instrument$prices()

Instrument$print()

Instrument$ticks()

TradeableInstrument$add_to_position()

TradeableInstrument$buy_at_best_bid_price()

TradeableInstrument$buy_at_best_offer_price()

TradeableInstrument$buy_at_fifth_best_bid_price()

TradeableInstrument$buy_at_fifth_best_offer_price()

TradeableInstrument$buy_at_fourth_best_bid_price()

TradeableInstrument$buy_at_fourth_best_offer_price()

TradeableInstrument$buy_at_last_price()

TradeableInstrument$buy_at_limit_price()

TradeableInstrument$buy_at_market_price()

TradeableInstrument$buy_at_marketable_price()

TradeableInstrument$buy_at_mid_price()

TradeableInstrument$buy_at_second_best_bid_price()

TradeableInstrument$buy_at_second_best_offer_price()

TradeableInstrument$buy_at_third_best_bid_price()

TradeableInstrument$buy_at_third_best_offer_price()

TradeableInstrument$buy_at_volume_weighted_average_price()

TradeableInstrument$cancel_open_orders()

TradeableInstrument$cancel_order()

TradeableInstrument$cancel_parent()

TradeableInstrument$liquidate_all_positions()

TradeableInstrument$liquidate_position()

TradeableInstrument$modify_order()

TradeableInstrument$parent()

TradeableInstrument$parent_orders()

TradeableInstrument$parent_trades()

TradeableInstrument$place_order()

TradeableInstrument$reduce_position()

TradeableInstrument$sell_at_best_bid_price()

TradeableInstrument$sell_at_best_offer_price()

TradeableInstrument$sell_at_fifth_best_bid_price()

TradeableInstrument$sell_at_fifth_best_offer_price()

TradeableInstrument$sell_at_fourth_best_bid_price()

TradeableInstrument$sell_at_fourth_best_offer_price()

TradeableInstrument$sell_at_last_price()

TradeableInstrument$sell_at_limit_price()

TradeableInstrument$sell_at_market_price()

TradeableInstrument$sell_at_marketable_price()

TradeableInstrument$sell_at_mid_price()

TradeableInstrument$sell_at_second_best_bid_price()

TradeableInstrument$sell_at_second_best_offer_price()

TradeableInstrument$sell_at_third_best_bid_price()

TradeableInstrument$sell_at_third_best_offer_price()

TradeableInstrument$sell_at_volume_weighted_average_price()

Option$greeks()

Option$implied_volatility()

IndexOption$new()

Looks the option up in UBI and checks that it is an option on an index.

Usage

IndexOption$new(
  instrument_id = NULL,
  exchange = NULL,
  segment = NULL,
  symbol = NULL,
  underlying_symbol = NULL,
  expiry_date = NULL,
  strike_price = NULL,
  option_type = NULL,
  underlying = NULL,
  unified_broker_interface = NULL
)

Arguments

instrument_id

The character UUID of the option, or NULL to look it up by exchange, segment and identity fields.

exchange

The character exchange, such as "nse", or NULL when instrument_id is given.

segment

The character segment, such as "equity_index_options", or NULL when instrument_id is given.

symbol

The character symbol, or NULL.

underlying_symbol

The character symbol of the index, such as "NIFTY", or NULL.

expiry_date

The expiry as a Date or a "YYYY-MM-DD" character value, or NULL.

strike_price

The numeric strike price, or NULL.

option_type

The character option type, "CE" or "PE", or NULL.

underlying

The Instrument the option is written on, or NULL to use UBI's link or the family's default.

unified_broker_interface

The UnifiedBrokerInterface to send requests through, or NULL to share one client among all instruments.

Details

Errors: signals TypeError when underlying is given and is not an Instrument; IndexOptionError when the option is written on something other than an index; OptionError when the instrument is a future rather than an option, or UBI gave it no strike price or option type; DerivativeError when the instrument is not a contract at all, or has no expiry date; TradeableInstrumentError when the instrument is an index; InstrumentError when UBI has no instrument matching the lookup; BadRequestError when the lookup is incomplete or malformed; and another UnifiedBrokerInterfaceError subclass for any other failure reported by, or on the way to, UBI.

Returns

A new IndexOption object.

IndexOption$clone()

The objects of this class are cloneable with this method.

Usage

IndexOption$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.